+++ lib/main.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
 import 'router/root_router_delegate.dart';
-import 'style/app_theme.dart';
-import 'utils/extensions/context_extensions.dart';
+import 'support/extensions/context_extensions.dart';
+import 'support/style/app_theme.dart';
 
 void main() {
   runApp(const MyApp());
