+++ lib/support/exceptions/api_exception.dart
