+++ lib/support/exceptions/local_database_exception.dart
