+++ lib/support/style/app_colors.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 
 abstract class AppColors {
   static const Color purple = Color.fromRGBO(40, 6, 99, 1);
-  static const Color deepPurple = Color.fromARGB(255, 16, 2, 40);
+  static const Color deepPurple = Color.fromRGBO(16, 2, 40, 1);
   static const Color lightPurple = Color.fromRGBO(246, 244, 251, 1);
   static const Color black = Color.fromRGBO(20, 20, 20, 1);
 }
